BinaryNetPytorch/requirements.txt
matplotlib==2.2.2
ipdb==0.11
numpy==1.14.0
pandas==0.22.0
seaborn==0.8.1
torch==0.3.1.post2
functional==0.7.0
torchvision==0.2.1
PyYAML==3.12
